move auth options to infra add github and google oauth providers
